Pushed settlement mechanism
