+++ _extractor/extractor.py
@@ -41,7 +41,7 @@ class Extractor():
         for row in tab:
             data.append(row)
 
-        return json.dumps(data[0])
+        return json.loads(json.dumps(data[0]))
 
 agent_id = 'egauge19821'
 db_path1 = 'postgresql://postgres:postgres@stargate/remote_agent_test_np'
@@ -49,4 +49,4 @@ sim_type = 'training'
 extractor = Extractor(db_path1)
 dataframe = extractor.from_metrics(0, sim_type, agent_id)
 
-print(dataframe.head())
+# print(dataframe.head())
